requirements.txt
django==5.0
django-filter==23.5
django-storages==1.14.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-rest-passwordreset==1.3.0
djangorestframework-simplejwt==5.3.1
django-import-export==3.3.4
django-ckeditor==6.7.0
drf-yasg==1.21.7
boto3==1.33.12
python-decouple==3.8
pillow==10.1.0
psycopg2-binary==2.9.9